config/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.11
Faker==20.1.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
kombu==5.6.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
